Completar entrevista por parte del administrador
